+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 ## Unreleased - 2020-xx-xx
 
+* Implement Logger middleware regex exclude pattern [#1723]
+[#1723]: https://github.com/actix/actix-web/pull/1723
+
 ### Fixed
 
 * awc now uses `Accept-Encoding: identity` instead of `Accept-Encoding: br` when no compression feature is enabled